Add files via upload
This file contains the yield dataset and have been checked for units in CY. Comments have been made in cells of every metanalytical study. The comments can be found in the cells: D5, D29, D30, D35, D49, D50, D51, D52, D154, D190, D194, D204 and in the sheet with the name: studies archive.